rename folder of evaluators to avoid name collisions
