+++ src/evaluate/evaluators/image_classification.py
