+++ src/evaluate/evaluators/token_classification.py
@@ -17,7 +17,7 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 from datasets import ClassLabel, Dataset, Sequence
 from typing_extensions import Literal
 
-from evaluate.evaluator.utils import DatasetColumn
+from evaluate.evaluators.utils import DatasetColumn
 
 from ..module import EvaluationModule
 from ..utils.file_utils import add_end_docstrings, add_start_docstrings
